// packages/video-player/typings/VideoPlayerProps.d.ts
/**
 * This file was generated from VideoPlayer.xml
 * WARNING: All changes made to this file will be overwritten
 * @author Mendix Widgets Team
 */
import { DynamicValue } from "mendix/properties";

interface CommonProps<Style> {
    style: Style[];
}

export interface VideoPlayerProps<Style> extends CommonProps<Style> {
    videoUrl: DynamicValue<string>;
    autoStart: boolean;
    muted: boolean;
    loop: boolean;
    aspectRatio: boolean;
    showControls: boolean;
}